+++ metric/internal/global/instruments.go
@@ -21,26 +21,25 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/embedded"
-	"go.opentelemetry.io/otel/metric/instrument"
 )
 
 // unwrapper unwraps to return the underlying instrument implementation.
 type unwrapper interface {
-	Unwrap() instrument.Observable
+	Unwrap() metric.Observable
 }
 
 type afCounter struct {
 	embedded.Float64ObservableCounter
-	instrument.Float64Observable
+	metric.Float64Observable
 
 	name string
-	opts []instrument.Float64ObservableCounterOption
+	opts []metric.Float64ObservableCounterOption
 
-	delegate atomic.Value //instrument.Float64ObservableCounter
+	delegate atomic.Value //metric.Float64ObservableCounter
 }
 
 var _ unwrapper = (*afCounter)(nil)
-var _ instrument.Float64ObservableCounter = (*afCounter)(nil)
+var _ metric.Float64ObservableCounter = (*afCounter)(nil)
 
 func (i *afCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableCounter(i.name, i.opts...)
@@ -51,25 +50,25 @@ func (i *afCounter) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *afCounter) Unwrap() instrument.Observable {
+func (i *afCounter) Unwrap() metric.Observable {
 	if ctr := i.delegate.Load(); ctr != nil {
-		return ctr.(instrument.Float64ObservableCounter)
+		return ctr.(metric.Float64ObservableCounter)
 	}
 	return nil
 }
 
 type afUpDownCounter struct {
 	embedded.Float64ObservableUpDownCounter
-	instrument.Float64Observable
+	metric.Float64Observable
 
 	name string
-	opts []instrument.Float64ObservableUpDownCounterOption
+	opts []metric.Float64ObservableUpDownCounterOption
 
-	delegate atomic.Value //instrument.Float64ObservableUpDownCounter
+	delegate atomic.Value //metric.Float64ObservableUpDownCounter
 }
 
 var _ unwrapper = (*afUpDownCounter)(nil)
-var _ instrument.Float64ObservableUpDownCounter = (*afUpDownCounter)(nil)
+var _ metric.Float64ObservableUpDownCounter = (*afUpDownCounter)(nil)
 
 func (i *afUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableUpDownCounter(i.name, i.opts...)
@@ -80,25 +79,25 @@ func (i *afUpDownCounter) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *afUpDownCounter) Unwrap() instrument.Observable {
+func (i *afUpDownCounter) Unwrap() metric.Observable {
 	if ctr := i.delegate.Load(); ctr != nil {
-		return ctr.(instrument.Float64ObservableUpDownCounter)
+		return ctr.(metric.Float64ObservableUpDownCounter)
 	}
 	return nil
 }
 
 type afGauge struct {
 	embedded.Float64ObservableGauge
-	instrument.Float64Observable
+	metric.Float64Observable
 
 	name string
-	opts []instrument.Float64ObservableGaugeOption
+	opts []metric.Float64ObservableGaugeOption
 
-	delegate atomic.Value //instrument.Float64ObservableGauge
+	delegate atomic.Value //metric.Float64ObservableGauge
 }
 
 var _ unwrapper = (*afGauge)(nil)
-var _ instrument.Float64ObservableGauge = (*afGauge)(nil)
+var _ metric.Float64ObservableGauge = (*afGauge)(nil)
 
 func (i *afGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableGauge(i.name, i.opts...)
@@ -109,25 +108,25 @@ func (i *afGauge) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *afGauge) Unwrap() instrument.Observable {
+func (i *afGauge) Unwrap() metric.Observable {
 	if ctr := i.delegate.Load(); ctr != nil {
-		return ctr.(instrument.Float64ObservableGauge)
+		return ctr.(metric.Float64ObservableGauge)
 	}
 	return nil
 }
 
 type aiCounter struct {
 	embedded.Int64ObservableCounter
-	instrument.Int64Observable
+	metric.Int64Observable
 
 	name string
-	opts []instrument.Int64ObservableCounterOption
+	opts []metric.Int64ObservableCounterOption
 
-	delegate atomic.Value //instrument.Int64ObservableCounter
+	delegate atomic.Value //metric.Int64ObservableCounter
 }
 
 var _ unwrapper = (*aiCounter)(nil)
-var _ instrument.Int64ObservableCounter = (*aiCounter)(nil)
+var _ metric.Int64ObservableCounter = (*aiCounter)(nil)
 
 func (i *aiCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableCounter(i.name, i.opts...)
@@ -138,25 +137,25 @@ func (i *aiCounter) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *aiCounter) Unwrap() instrument.Observable {
+func (i *aiCounter) Unwrap() metric.Observable {
 	if ctr := i.delegate.Load(); ctr != nil {
-		return ctr.(instrument.Int64ObservableCounter)
+		return ctr.(metric.Int64ObservableCounter)
 	}
 	return nil
 }
 
 type aiUpDownCounter struct {
 	embedded.Int64ObservableUpDownCounter
-	instrument.Int64Observable
+	metric.Int64Observable
 
 	name string
-	opts []instrument.Int64ObservableUpDownCounterOption
+	opts []metric.Int64ObservableUpDownCounterOption
 
-	delegate atomic.Value //instrument.Int64ObservableUpDownCounter
+	delegate atomic.Value //metric.Int64ObservableUpDownCounter
 }
 
 var _ unwrapper = (*aiUpDownCounter)(nil)
-var _ instrument.Int64ObservableUpDownCounter = (*aiUpDownCounter)(nil)
+var _ metric.Int64ObservableUpDownCounter = (*aiUpDownCounter)(nil)
 
 func (i *aiUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableUpDownCounter(i.name, i.opts...)
@@ -167,25 +166,25 @@ func (i *aiUpDownCounter) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *aiUpDownCounter) Unwrap() instrument.Observable {
+func (i *aiUpDownCounter) Unwrap() metric.Observable {
 	if ctr := i.delegate.Load(); ctr != nil {
-		return ctr.(instrument.Int64ObservableUpDownCounter)
+		return ctr.(metric.Int64ObservableUpDownCounter)
 	}
 	return nil
 }
 
 type aiGauge struct {
 	embedded.Int64ObservableGauge
-	instrument.Int64Observable
+	metric.Int64Observable
 
 	name string
-	opts []instrument.Int64ObservableGaugeOption
+	opts []metric.Int64ObservableGaugeOption
 
-	delegate atomic.Value //instrument.Int64ObservableGauge
+	delegate atomic.Value //metric.Int64ObservableGauge
 }
 
 var _ unwrapper = (*aiGauge)(nil)
-var _ instrument.Int64ObservableGauge = (*aiGauge)(nil)
+var _ metric.Int64ObservableGauge = (*aiGauge)(nil)
 
 func (i *aiGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableGauge(i.name, i.opts...)
@@ -196,9 +195,9 @@ func (i *aiGauge) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *aiGauge) Unwrap() instrument.Observable {
+func (i *aiGauge) Unwrap() metric.Observable {
 	if ctr := i.delegate.Load(); ctr != nil {
-		return ctr.(instrument.Int64ObservableGauge)
+		return ctr.(metric.Int64ObservableGauge)
 	}
 	return nil
 }
@@ -208,12 +207,12 @@ type sfCounter struct {
 	embedded.Float64Counter
 
 	name string
-	opts []instrument.Float64CounterOption
+	opts []metric.Float64CounterOption
 
-	delegate atomic.Value //instrument.Float64Counter
+	delegate atomic.Value //metric.Float64Counter
 }
 
-var _ instrument.Float64Counter = (*sfCounter)(nil)
+var _ metric.Float64Counter = (*sfCounter)(nil)
 
 func (i *sfCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Counter(i.name, i.opts...)
@@ -224,9 +223,9 @@ func (i *sfCounter) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *sfCounter) Add(ctx context.Context, incr float64, opts ...instrument.AddOption) {
+func (i *sfCounter) Add(ctx context.Context, incr float64, opts ...metric.AddOption) {
 	if ctr := i.delegate.Load(); ctr != nil {
-		ctr.(instrument.Float64Counter).Add(ctx, incr, opts...)
+		ctr.(metric.Float64Counter).Add(ctx, incr, opts...)
 	}
 }
 
@@ -234,12 +233,12 @@ type sfUpDownCounter struct {
 	embedded.Float64UpDownCounter
 
 	name string
-	opts []instrument.Float64UpDownCounterOption
+	opts []metric.Float64UpDownCounterOption
 
-	delegate atomic.Value //instrument.Float64UpDownCounter
+	delegate atomic.Value //metric.Float64UpDownCounter
 }
 
-var _ instrument.Float64UpDownCounter = (*sfUpDownCounter)(nil)
+var _ metric.Float64UpDownCounter = (*sfUpDownCounter)(nil)
 
 func (i *sfUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64UpDownCounter(i.name, i.opts...)
@@ -250,9 +249,9 @@ func (i *sfUpDownCounter) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *sfUpDownCounter) Add(ctx context.Context, incr float64, opts ...instrument.AddOption) {
+func (i *sfUpDownCounter) Add(ctx context.Context, incr float64, opts ...metric.AddOption) {
 	if ctr := i.delegate.Load(); ctr != nil {
-		ctr.(instrument.Float64UpDownCounter).Add(ctx, incr, opts...)
+		ctr.(metric.Float64UpDownCounter).Add(ctx, incr, opts...)
 	}
 }
 
@@ -260,12 +259,12 @@ type sfHistogram struct {
 	embedded.Float64Histogram
 
 	name string
-	opts []instrument.Float64HistogramOption
+	opts []metric.Float64HistogramOption
 
-	delegate atomic.Value //instrument.Float64Histogram
+	delegate atomic.Value //metric.Float64Histogram
 }
 
-var _ instrument.Float64Histogram = (*sfHistogram)(nil)
+var _ metric.Float64Histogram = (*sfHistogram)(nil)
 
 func (i *sfHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Histogram(i.name, i.opts...)
@@ -276,9 +275,9 @@ func (i *sfHistogram) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *sfHistogram) Record(ctx context.Context, x float64, opts ...instrument.RecordOption) {
+func (i *sfHistogram) Record(ctx context.Context, x float64, opts ...metric.RecordOption) {
 	if ctr := i.delegate.Load(); ctr != nil {
-		ctr.(instrument.Float64Histogram).Record(ctx, x, opts...)
+		ctr.(metric.Float64Histogram).Record(ctx, x, opts...)
 	}
 }
 
@@ -286,12 +285,12 @@ type siCounter struct {
 	embedded.Int64Counter
 
 	name string
-	opts []instrument.Int64CounterOption
+	opts []metric.Int64CounterOption
 
-	delegate atomic.Value //instrument.Int64Counter
+	delegate atomic.Value //metric.Int64Counter
 }
 
-var _ instrument.Int64Counter = (*siCounter)(nil)
+var _ metric.Int64Counter = (*siCounter)(nil)
 
 func (i *siCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Counter(i.name, i.opts...)
@@ -302,9 +301,9 @@ func (i *siCounter) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *siCounter) Add(ctx context.Context, x int64, opts ...instrument.AddOption) {
+func (i *siCounter) Add(ctx context.Context, x int64, opts ...metric.AddOption) {
 	if ctr := i.delegate.Load(); ctr != nil {
-		ctr.(instrument.Int64Counter).Add(ctx, x, opts...)
+		ctr.(metric.Int64Counter).Add(ctx, x, opts...)
 	}
 }
 
@@ -312,12 +311,12 @@ type siUpDownCounter struct {
 	embedded.Int64UpDownCounter
 
 	name string
-	opts []instrument.Int64UpDownCounterOption
+	opts []metric.Int64UpDownCounterOption
 
-	delegate atomic.Value //instrument.Int64UpDownCounter
+	delegate atomic.Value //metric.Int64UpDownCounter
 }
 
-var _ instrument.Int64UpDownCounter = (*siUpDownCounter)(nil)
+var _ metric.Int64UpDownCounter = (*siUpDownCounter)(nil)
 
 func (i *siUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64UpDownCounter(i.name, i.opts...)
@@ -328,9 +327,9 @@ func (i *siUpDownCounter) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *siUpDownCounter) Add(ctx context.Context, x int64, opts ...instrument.AddOption) {
+func (i *siUpDownCounter) Add(ctx context.Context, x int64, opts ...metric.AddOption) {
 	if ctr := i.delegate.Load(); ctr != nil {
-		ctr.(instrument.Int64UpDownCounter).Add(ctx, x, opts...)
+		ctr.(metric.Int64UpDownCounter).Add(ctx, x, opts...)
 	}
 }
 
@@ -338,12 +337,12 @@ type siHistogram struct {
 	embedded.Int64Histogram
 
 	name string
-	opts []instrument.Int64HistogramOption
+	opts []metric.Int64HistogramOption
 
-	delegate atomic.Value //instrument.Int64Histogram
+	delegate atomic.Value //metric.Int64Histogram
 }
 
-var _ instrument.Int64Histogram = (*siHistogram)(nil)
+var _ metric.Int64Histogram = (*siHistogram)(nil)
 
 func (i *siHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Histogram(i.name, i.opts...)
@@ -354,8 +353,8 @@ func (i *siHistogram) setDelegate(m metric.Meter) {
 	i.delegate.Store(ctr)
 }
 
-func (i *siHistogram) Record(ctx context.Context, x int64, opts ...instrument.RecordOption) {
+func (i *siHistogram) Record(ctx context.Context, x int64, opts ...metric.RecordOption) {
 	if ctr := i.delegate.Load(); ctr != nil {
-		ctr.(instrument.Int64Histogram).Record(ctx, x, opts...)
+		ctr.(metric.Int64Histogram).Record(ctx, x, opts...)
 	}
 }